Move linkTypes to common config
